requirements.txt
 # Core Libraries
 streamlit==1.32.0
 python-docx==0.8.11
 torch==2.1.0
 torchvision==0.16.0
 transformers==4.36.0
 accelerate==0.24.0
 datasets==2.14.6
 opencv-python==4.8.1.78
 easyocr==1.7.2
 pytesseract==0.3.13
 layoutparser[layoutlm]==0.3.4
#  layoutlm==0.0.1
 Pillow==10.2.0
 numpy==1.24.3
 pandas==2.1.3
 matplotlib==3.8.2
 seaborn==0.13.0
 plotly==5.17.0
 scikit-learn==1.4.0
 nltk==3.8.1
 rouge-score==0.1.2
 sacrebleu==2.3.1
 openai==1.3.3
 groq==0.30.0  # downloaded epxlicitly version 0.30.0
 requests==2.31.0
 PyYAML==6.0.1
 tqdm==4.66.1
#  logging==0.4.9.6
 json5==0.9.14
 boto3==1.29.7
 sacremoses==0.1.1
 nibabel==5.3.2
 pydicom==2.4.4